Feature/implement login and register

+++ src/main/java/org/pawpal/service/PetService.java
@@ -7,9 +7,11 @@ import org.pawpal.repository.PetRepository;
 import org.pawpal.repository.UserRepository;
 import org.pawpal.util.MapperUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 
+import java.io.IOException;
 import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.Objects;
@@ -23,13 +25,8 @@ public class PetService {
     @Autowired
     private UserRepository userRepository;
 
-    public List<PetDTO> getAllPets() {
-        return petRepository.findAll().stream()
-                .map(MapperUtil::toPetDTO)
-                .toList();
-    }
-
-    public List<PetDTO> getPetsByUserEmail(String email) {
+    public List<PetDTO> getPetsByUserEmail() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
         return petRepository.findAll().stream()
                 .filter(p -> Objects.equals(p.getOwner().getEmail(), email))
                 .map(MapperUtil::toPetDTO)
@@ -42,8 +39,9 @@ public class PetService {
         } else throw new RuntimeException("Pet not found");
     }
 
-    public PetDTO createPet(PetDTO petDTO) {
-        User user = userRepository.findByEmail(petDTO.getEmail());
+    public PetDTO createPet(PetDTO petDTO) throws IOException {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findByEmail(email).get();
         Pet pet = MapperUtil.toPet(petDTO, user);
         return MapperUtil.toPetDTO(petRepository.save(pet));
     }
@@ -55,7 +53,8 @@ public class PetService {
             existingPet.setAge(petDTO.getAge());
             existingPet.setBreed(petDTO.getBreed());
             existingPet.setWeight(petDTO.getWeight());
-            existingPet.setMedicalHistory(petDTO.getMedicalHistory());
+            existingPet.setGender(petDTO.isGender());
+            existingPet.setOwner(userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName()).get());
             Pet updatedPet = petRepository.save(existingPet);
             return MapperUtil.toPetDTO(updatedPet);
         } else throw new RuntimeException("Pet not found");
@@ -63,7 +62,13 @@ public class PetService {
 
     public void deletePet(Long id)  {
         if (petRepository.existsById(id)) {
-            petRepository.deleteById(id);
+            String email = SecurityContextHolder.getContext().getAuthentication().getName();
+            //User user = userRepository.findByEmail(email).get();
+            Pet pet = petRepository.findById(id).get();
+            if(pet.getOwner().getEmail().equals(email))
+                petRepository.deleteById(id);
+            else
+                throw new RuntimeException("You are not authorized to delete this pet");
         } else throw new RuntimeException("Pet not found");
     }
